these variables are signed not unsigned so comparison was failing allowing a -1 index lookup on array
git-svn-id: svn://fly.thruhere.net/forsaken@2456 1361c5d3-c82f-4d94-b015-8aedb5f389d9

+++ ProjectX/trunk/src/bgobjects.c
@@ -781,7 +781,7 @@ void ChangeBGState( BGOBJECT * Object, u_int16_t OwnerType, u_int16_t Owner, int
 				{
 					if( ( OwnerType == OWNER_SHIP ) && ( Owner == WhoIAm ) )
 					{
-						if( Object->PickupNeeded != (u_int16_t) -1 )
+						if( Object->PickupNeeded != -1 )
 						{
 							if( PickupsGot[ Object->PickupNeeded ] )
 							{
@@ -810,7 +810,7 @@ void ChangeBGState( BGOBJECT * Object, u_int16_t OwnerType, u_int16_t Owner, int
 				{
 					if( ( OwnerType == OWNER_SHIP ) && ( Owner == WhoIAm ) )
 					{
-						if( Object->PickupNeeded != (u_int16_t) -1 )
+						if( Object->PickupNeeded != -1 )
 						{
 							if( PickupsGot[ Object->PickupNeeded ] )
 							{
